Update build and layout.
Re-introduce Python dependencies on ziglang for easier bootstrap. Now
all you need is `uv` to develop the project. Its still probably best
to have zig installed on your own; preferably using something like
anyzig.

Also move the libzip test-bed `zipopen` executable to a new top-level
`bench/` directory. Its not a proper benchmark, but that is the general
intent and how it was used in practice when evaluating libzip & zstd.